Feature/scrum 130 implement http live streaming

+++ src/Yuviron.Application/Features/Admin/Albums/Queries/GetAlbumById/AlbumDetailsDto.cs
@@ -13,7 +13,7 @@ public record AlbumDetailsDto(
     DateTime? ScheduledPublishAt,
     int TracksCount,     
     long TotalDurationMs, //mm:ss
-    int TotalPlays,     
+    long TotalPlays,     
     DateTime CreatedAt,  
     DateTime UpdatedAt, 
     List<ArtistSimpleDto> Artists
